Update blackr extension
- Merge branch \'contributions/merge-1782336220984\'
- Pull contributions
- feat(macos): rewrite overlay in Swift - Add changelog entry for version 1.2.0 - Bump package version to 1.2.0 - Replace macOS overlay implementation with Swift
- fix(build): remove obsolete native overlay code and simplify scripts
