+++ extensions/blackr/.gitignore
@@ -14,10 +14,11 @@ raycast-env.d.ts
 compiled_raycast_swift
 compiled_raycast_rust
 
+# Swift build artifacts
+.build
+
 # misc
 .DS_Store
 
 # project-specific
-assets/blackr-overlay-arm64
-assets/blackr-overlay-x64
 .codex
