+++ extensions/blackr/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Blackr Changelog
 
+## [1.2.0] - {PR_MERGE_DATE}
+
+- Update the macOS overlay to use Swift.
+
 ## [1.1.0] - 2026-06-22
 
 - Add support for Raycast on Windows.
